.
This repository contains a clone of the popular e-commerce website Myntra. The project aims to replicate the core functionalities and design of Myntra, providing a comprehensive learning experience in web development and showcasing skills in front-end and back-end technologies.

Contributing
Contributions are welcome! If you have any ideas, suggestions, or improvements, please feel free to submit a pull request or open an issue.